requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
charset-normalizer==2.0.7; python_version >= '3'
click==8.0.3; python_version >= '3.6'
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django==3.2.9
djangorestframework==3.12.4
filelock==3.3.2; python_version >= '3.6'
huggingface-hub==0.1.2; python_version >= '3.6'
idna==3.3; python_version >= '3'
joblib==1.1.0; python_version >= '3.6'
numpy==1.21.4; python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'
packaging==21.2; python_version >= '3.6'
pandas==1.3.4
pillow==8.4.0; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
regex==2021.11.2
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sacremoses==0.0.46
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tokenizers==0.10.3
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
